release into the public domain
Note, this repository contains an implementation of a Smacker
video and audio decoder based on the documentation of the Smacker
file format at [1]. RAD Tools [2] still holds copyright to the
original Smacker implementation.

That being said, any code in this repository is dedicated into
the public domain. It may only be used for non-commercial projects,
in the same manner as libsmacker [3].

[1]: https://wiki.multimedia.cx/index.php?title=Smacker
[2]: http://www.radgametools.com/
[3]: http://libsmacker.sourceforge.net/#license